requirements.txt
jpype1
jupyter
notebook
tabula-py
